update server port to 3005 and enhance data retrieval with optional hours filter

+++ src/controllers/sensorController.js
@@ -6,7 +6,7 @@ const getAllData = async (req, res) => {
     let filter = {};
 
     if (hours) {
-      const parsedHours = parseInt(hours);
+      const parsedHours = parseFloat(hours); // <-- ubah ke float
       if (!isNaN(parsedHours)) {
         // cari data terbaru
         const latestData = await SensorData.findOne().sort({ timestamp: -1 });
